infra/requirements.txt
aws-cdk-lib==2.151.0
constructs>=10.0.0,<11.0.0
boto3==1.28.51
botocore==1.31.51